Removed fist fighting code and added projectiles instead. Enemies take damage from these projectiles.

+++ Adam/Adam/Misc/TimeFreeze.cs
@@ -19,7 +19,7 @@ namespace Adam.Misc
             _isTimeFrozen = false;
         }
 
-        public void AddFrozenTime(double time)
+        public void AddFrozenTime(int time)
         {
             _isTimeFrozen = true;
             _freezeTimer.ResetAndWaitFor(time);
